XTENLANG-2075 enable x10.core.Int/Byte/Short/Long/Float/Double
* implicit un/boxing through type conversion replaced
  with explicit un/boxing through $box()/$unbox()
* isBoxedType() is used instead of explicity conditions
* x10.core numerics are reboxed to java.lang numerics on String.format()
* Arithmetic, Ordered etc. are implemented for all numeric types
  (fix for XTENLANG-2075)



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@22214 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-java/x10/core/Double.java
@@ -24,7 +24,7 @@ final public class Double extends Numeric implements java.lang.Comparable<Double
 {
     private static final long serialVersionUID = 1L;
     
-    public static final x10.rtt.RuntimeType<?> $RTT = Types.INT;
+    public static final x10.rtt.RuntimeType<?> $RTT = Types.DOUBLE;
     public x10.rtt.RuntimeType<?> $getRTT() {return $RTT;}
     public x10.rtt.Type<?> $getParam(int i) {return null;}
