+++ x10.runtime/src-java/x10/core/Float.java
@@ -24,7 +24,7 @@ final public class Float extends Numeric implements java.lang.Comparable<Float>,
 {
     private static final long serialVersionUID = 1L;
     
-    public static final x10.rtt.RuntimeType<?> $RTT = Types.INT;
+    public static final x10.rtt.RuntimeType<?> $RTT = Types.FLOAT;
     public x10.rtt.RuntimeType<?> $getRTT() {return $RTT;}
     public x10.rtt.Type<?> $getParam(int i) {return null;}
